+++ src/core/mustache/mod.rs
@@ -2,5 +2,5 @@ mod eval;
 mod model;
 mod parse;
 
-pub use eval::{Eval, ValueStringEval};
+pub use eval::Eval;
 pub use model::*;
